requirements.txt
flask
line-bot-sdk
openai==0.28.1
python-dotenv
